backend/requirements.txt
alembic==1.17.1
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
distro==1.9.0
dnspython==2.8.0
dotenv==0.9.9
email-validator==2.3.0
fastapi==0.121.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.186.0
google-auth==2.42.1
google-auth-httplib2==0.2.1
google-genai==1.49.0
googleapis-common-protos==1.71.0
greenlet==3.2.4
groq==0.33.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
jiter==0.12.0
Mako==1.3.10
MarkupSafe==3.0.3
openai==2.7.1
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.11
pwdlib==0.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
pyparsing==3.2.5
python-dotenv==1.2.1
python-multipart==0.0.20
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlmodel==0.0.27
starlette==0.49.3
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
websockets==15.0.1
